feat: adjustment to not-found
